test: slim runtime IT stabilization to safety nets + shared query helper
Drop the one-method-per-class splits and keep only the hardening that pays
for itself.

RuntimeMigrationAbstractTest carries @DirtiesContext(AFTER_CLASS) (via
AbstractMigratorTest), so the Spring context and the CPT-managed Zeebe cluster
are torn down after every test class. Splitting an N-method class into N
one-method classes therefore triggers N full cluster bootstraps instead of one,
and each bootstrap is exactly the "Partition unhealthy / Services not installed"
window that the flakiness comes from -- so the split made the suite slower and
more exposed, not more stable. The actual red CI
(ValidationJobTypeDisabledTest -> "Failed to activate jobs for type: migrator")
is in a class that was never split and is the activateJobs transient-timeout
crash fixed by the production retry in #1488, not anything the splits address.

- Fold PageSizeMultiLevel, SeparateJobTypes and DeclarativeConfiguration split
  classes back into single classes; delete the two split-out files
- Keep the useful hardening: class-level @Timeout, bounded client search
  .join(10s), 30s Awaitility bound, timeout-tolerant cleanup, the merged
  duplicate pagination scenario, and the redundant String.format removal
- Extract the bounded-search / timeout-detection logic that was duplicated
  verbatim across RuntimeMigrationExtension and RuntimeMigrationAbstractTest
  into a single BoundedCamundaQueries helper
- Document the 1-2 occurrence skip-log race in ExternalTrafficTest

Real fix for the activateJobs timeout lives in #1488; this PR depends on it and
should be rebased once #1488 merges.
