+++ data-migrator/qa/integration-tests/src/test/java/io/camunda/migration/data/qa/extension/RuntimeMigrationExtension.java
@@ -18,11 +18,10 @@ import io.camunda.client.api.search.response.ProcessInstance;
 import io.camunda.client.api.search.response.Variable;
 import io.camunda.migration.data.RuntimeMigrator;
 import io.camunda.migration.data.impl.clients.DbClient;
-import java.net.SocketTimeoutException;
+import io.camunda.migration.data.qa.util.BoundedCamundaQueries;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 import org.awaitility.Awaitility;
 import org.camunda.bpm.engine.RepositoryService;
 import org.camunda.bpm.engine.RuntimeService;
@@ -52,8 +51,6 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
  */
 @Component
 public class RuntimeMigrationExtension implements AfterEachCallback, ApplicationContextAware {
-
-  private static final long CAMUNDA_CLIENT_REQUEST_TIMEOUT_SECONDS = 10;
 
   private static ApplicationContext applicationContext;
 
@@ -113,7 +110,7 @@ public class RuntimeMigrationExtension implements AfterEachCallback, Application
     CamundaClient camundaClient = getCamundaClientBean();
     if (camundaClient != null) {
       // C8
-      List<ProcessInstance> items = searchProcessInstancesForCleanup(camundaClient);
+      List<ProcessInstance> items = BoundedCamundaQueries.searchProcessInstancesForCleanup(camundaClient);
       for (ProcessInstance processInstance : items) {
         try {
           camundaClient.newDeleteResourceCommand(processInstance.getProcessInstanceKey()).execute();
@@ -170,7 +167,7 @@ public class RuntimeMigrationExtension implements AfterEachCallback, Application
     if (camundaClient == null) {
       return Optional.empty();
     }
-    List<Variable> variables = searchVariables(camundaClient);
+    List<Variable> variables = BoundedCamundaQueries.searchVariables(camundaClient);
 
     return variables.stream()
         .filter(v -> v.getProcessInstanceKey().equals(processInstanceKey))
@@ -188,44 +185,8 @@ public class RuntimeMigrationExtension implements AfterEachCallback, Application
         .atMost(30, TimeUnit.SECONDS)
         .ignoreException(ClientException.class)
         .untilAsserted(() -> {
-          assertThat(searchProcessInstances(camundaClient)).hasSize(expected);
+          assertThat(BoundedCamundaQueries.searchProcessInstances(camundaClient)).hasSize(expected);
         });
-  }
-
-  private List<ProcessInstance> searchProcessInstances(CamundaClient camundaClient) {
-    return camundaClient.newProcessInstanceSearchRequest()
-        .send()
-        .join(CAMUNDA_CLIENT_REQUEST_TIMEOUT_SECONDS, TimeUnit.SECONDS)
-        .items();
-  }
-
-  private List<ProcessInstance> searchProcessInstancesForCleanup(CamundaClient camundaClient) {
-    try {
-      return searchProcessInstances(camundaClient);
-    } catch (ClientException e) {
-      if (isCausedByTimeout(e)) {
-        return List.of();
-      }
-      throw e;
-    }
-  }
-
-  private List<Variable> searchVariables(CamundaClient camundaClient) {
-    return camundaClient.newVariableSearchRequest()
-        .send()
-        .join(CAMUNDA_CLIENT_REQUEST_TIMEOUT_SECONDS, TimeUnit.SECONDS)
-        .items();
-  }
-
-  private boolean isCausedByTimeout(Throwable exception) {
-    Throwable cause = exception;
-    while (cause != null) {
-      if (cause instanceof TimeoutException || cause instanceof SocketTimeoutException) {
-        return true;
-      }
-      cause = cause.getCause();
-    }
-    return false;
   }
 }
