+++ data-migrator/qa/integration-tests/src/test/java/io/camunda/migration/data/qa/runtime/RuntimeMigrationAbstractTest.java
@@ -19,12 +19,11 @@ import io.camunda.client.api.search.response.Variable;
 import io.camunda.migration.data.RuntimeMigrator;
 import io.camunda.migration.data.impl.clients.DbClient;
 import io.camunda.migration.data.qa.AbstractMigratorTest;
+import io.camunda.migration.data.qa.util.BoundedCamundaQueries;
 import io.camunda.process.test.api.CamundaSpringProcessTest;
-import java.net.SocketTimeoutException;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 import org.awaitility.Awaitility;
 import org.camunda.bpm.engine.RepositoryService;
 import org.camunda.bpm.engine.RuntimeService;
@@ -35,8 +34,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 @CamundaSpringProcessTest
 public abstract class RuntimeMigrationAbstractTest extends AbstractMigratorTest {
-
-  private static final long CAMUNDA_CLIENT_REQUEST_TIMEOUT_SECONDS = 10;
 
   // Migrator ---------------------------------------
 
@@ -70,7 +67,7 @@ public abstract class RuntimeMigrationAbstractTest extends AbstractMigratorTest 
     repositoryService.createDeploymentQuery().list().forEach(d -> repositoryService.deleteDeployment(d.getId(), true));
 
     // C8
-    List<ProcessInstance> items = searchProcessInstancesForCleanup();
+    List<ProcessInstance> items = BoundedCamundaQueries.searchProcessInstancesForCleanup(camundaClient);
     for (ProcessInstance processInstance : items) {
       try {
         camundaClient.newDeleteResourceCommand(processInstance.getProcessInstanceKey()).execute();
@@ -88,7 +85,7 @@ public abstract class RuntimeMigrationAbstractTest extends AbstractMigratorTest 
   }
 
   protected Optional<Variable> getVariableByScope(Long processInstanceKey, Long scopeKey, String variableName) {
-    List<Variable> variables = searchVariables();
+    List<Variable> variables = BoundedCamundaQueries.searchVariables(camundaClient);
 
     return variables.stream()
         .filter(v -> v.getProcessInstanceKey().equals(processInstanceKey))
@@ -102,44 +99,8 @@ public abstract class RuntimeMigrationAbstractTest extends AbstractMigratorTest 
         .atMost(30, TimeUnit.SECONDS)
         .ignoreException(ClientException.class)
         .untilAsserted(() -> {
-          assertThat(searchProcessInstances()).hasSize(expected);
+          assertThat(BoundedCamundaQueries.searchProcessInstances(camundaClient)).hasSize(expected);
         });
-  }
-
-  private List<ProcessInstance> searchProcessInstances() {
-    return camundaClient.newProcessInstanceSearchRequest()
-        .send()
-        .join(CAMUNDA_CLIENT_REQUEST_TIMEOUT_SECONDS, TimeUnit.SECONDS)
-        .items();
-  }
-
-  private List<ProcessInstance> searchProcessInstancesForCleanup() {
-    try {
-      return searchProcessInstances();
-    } catch (ClientException e) {
-      if (isCausedByTimeout(e)) {
-        return List.of();
-      }
-      throw e;
-    }
-  }
-
-  private List<Variable> searchVariables() {
-    return camundaClient.newVariableSearchRequest()
-        .send()
-        .join(CAMUNDA_CLIENT_REQUEST_TIMEOUT_SECONDS, TimeUnit.SECONDS)
-        .items();
-  }
-
-  private boolean isCausedByTimeout(Throwable exception) {
-    Throwable cause = exception;
-    while (cause != null) {
-      if (cause instanceof TimeoutException || cause instanceof SocketTimeoutException) {
-        return true;
-      }
-      cause = cause.getCause();
-    }
-    return false;
   }
 
 }
